refactor: docs img path, simplify store
