+++ src/components/Documentation/Documentation.module.scss
@@ -58,7 +58,7 @@
 }
 .docs__error_img {
   margin-bottom: 2rem;
-  background-image: url('../../../public/png/Error.png');
+  background-image: url('/png/Error.png');
   background-size: cover;
   width: 3rem;
   height: 3rem;
