+++ src/hooks/redux-hooks.ts
@@ -3,7 +3,6 @@ import { bindActionCreators } from '@reduxjs/toolkit';
 
 import { RootState } from 'shared/store/store';
 
-import { endpointActions } from '../shared/store/reducers/endpointSlice';
 import { docsActions } from '../shared/store/slices/docsSlice';
 import {
   setHeaders,
@@ -11,6 +10,7 @@ import {
   setResponse,
   setVariables,
 } from '../shared/store/slices/editorSlice';
+import { endpointActions } from '../shared/store/slices/endpointSlice';
 
 const actions = {
   ...endpointActions,
