+++ src/shared/store/slices/docsSlice.ts
@@ -3,13 +3,11 @@ import { createSlice, PayloadAction } from '@reduxjs/toolkit';
 import GraphQLMethod from 'shared/enums/GraphQLMethod';
 
 interface DocsState {
-  isOpened: boolean;
   targetName: string;
   targetType: string;
   method: GraphQLMethod;
 }
 const initialState: DocsState = {
-  isOpened: false,
   targetName: 'index',
   targetType: '',
   method: GraphQLMethod.Query,
@@ -19,10 +17,6 @@ const docsSlice = createSlice({
   name: 'docs',
   initialState,
   reducers: {
-    setDocsIsOpened: (state, action: PayloadAction<boolean>) => {
-      state.isOpened = action.payload;
-      state.targetName = 'index';
-    },
     setDocsTargetName: (state, action: PayloadAction<string>) => {
       state.targetName = action.payload;
     },
