+++ src/shared/store/slices/endpointSlice.ts
